[AND-147] - Move Updates preferences to shared module
