+++ app-games/src/main/java/com/aptoide/android/aptoidegames/di/RepositoryModule.kt
@@ -39,7 +39,6 @@ import com.aptoide.android.aptoidegames.search.repository.AppGamesAutoCompleteSu
 import com.aptoide.android.aptoidegames.search.repository.AppGamesAutoCompleteSuggestionsRepository.AutoCompleteSearchRetrofitService
 import com.aptoide.android.aptoidegames.search.repository.AppGamesSearchStoreManager
 import com.aptoide.android.aptoidegames.themeDataStore
-import com.aptoide.android.aptoidegames.updatesPreferencesDataStore
 import com.aptoide.android.aptoidegames.userFeatureFlagsDataStore
 import dagger.Module
 import dagger.Provides
@@ -120,15 +119,6 @@ class RepositoryModule {
     @ApplicationContext appContext: Context,
   ): DataStore<Preferences> {
     return appContext.networkPreferencesDataStore
-  }
-
-  @Singleton
-  @Provides
-  @UpdatesPreferencesDataStore
-  fun provideUpdatesPreferencesDataStore(
-    @ApplicationContext appContext: Context,
-  ): DataStore<Preferences> {
-    return appContext.updatesPreferencesDataStore
   }
 
   @Singleton
@@ -251,10 +241,6 @@ annotation class PermissionsDataStore
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
 annotation class NetworkPreferencesDataStore
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class UpdatesPreferencesDataStore
 
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
