+++ feature_updates/build.gradle.kts
@@ -21,6 +21,9 @@ dependencies {
   ksp(LibraryDependency.ROOM_COMPILER)
   implementation(LibraryDependency.ROOM_KTX)
 
+  //Store
+  implementation(LibraryDependency.DATASTORE)
+
   //WorkManager
   implementation(LibraryDependency.WORK_MANAGER)
   implementation(LibraryDependency.HILT_WORK)
