+++ feature_updates/src/main/java/cm/aptoide/pt/feature_updates/repository/UpdatesPreferencesRepository.kt
@@ -1,10 +1,10 @@
-package com.aptoide.android.aptoidegames.updates.repository
+package cm.aptoide.pt.feature_updates.repository
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
-import com.aptoide.android.aptoidegames.di.UpdatesPreferencesDataStore
+import cm.aptoide.pt.feature_updates.di.UpdatesPreferencesDataStore
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
